third_party/blink/web_tests/transitions/repeated-firing-background-color-expected.txt

This should not be continually firing transitions
PASS: Number of active transitions is (0) as expected
